requirements.txt
# Web framework
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9

# Data validation
pydantic==2.8.2

# LangChain — top level only, let pip resolve core version
langchain==0.3.7
langchain-community==0.3.7
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langchain-chroma==0.1.4
langchain-groq==0.2.1

# Vector store
chromadb==0.5.23

# Embeddings
sentence-transformers==3.3.1

# Document processing
pypdf==4.3.1
beautifulsoup4==4.12.3
requests==2.32.3

# Async job processing
celery==5.3.6
redis==5.0.1

# Utilities
python-dotenv==1.0.1
numpy<2

google-cloud-storage==2.14.0